♻️ refactor: キーボードショートカットとエラーハンドリングの重複を解消
- use-keyboard-shortcuts.ts: app.tsx/community.tsx で同一だった34行を共通hookに抽出
- errors.ts: 9箇所で重複していた e instanceof Error パターンを getErrorMessage() に統合
